requirements.txt
Pillow==9.2.0
PyMuPDF==1.21.1
python-pptx==0.6.21
streamlit==1.16.0

